cpufreq: Force the LITTLE cluster to only use the default governor
Since the LITTLE cluster now only has two frequencies, both of which are
very close to one another, it doesn't make much sense to keep schedutil
running for it when there's almost no power difference between the two
frequencies. As such, force the LITTLE cluster to stick to the default
CPU governor, which is always sanely set to performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Hadid Garcia <hadid.garcia1@gmail.com>